requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-CORS==4.0.0
python-dotenv==1.0.0
openai==1.3.7
PyMySQL==1.1.0
cryptography==41.0.7